LLM/Protfolio/requirements.txt
flask
transformers
langchain
langchain-community
langchain-classic
langchain-openai
langchain-huggingface
sentence-transformers
chromadb
python-dotenv